+++ spike/riscv/insns/ss_ld_b.h
@@ -5,8 +5,8 @@ auto sizeReg = insn.uve_conf_size();
 auto strideReg = insn.uve_conf_stride();
 
 reg_t base = READ_REG(baseReg);
-int32_t size = READ_REG(sizeReg);
-int32_t stride = READ_REG(strideReg);
+size_t size = READ_REG(sizeReg);
+size_t stride = READ_REG(strideReg);
 
 P.SU.makeStreamRegister<std::uint8_t>(RegisterConfig::Load, streamReg);
 /*operateRegister(P.SU, streamReg, [=](auto& reg) {
