+++ spike/riscv/insns/ss_ld_h.h
@@ -5,8 +5,8 @@ auto sizeReg = insn.uve_conf_size();
 auto strideReg = insn.uve_conf_stride();
 
 reg_t base = READ_REG(baseReg);
-int32_t size = READ_REG(sizeReg);
-int32_t stride = READ_REG(strideReg);
+size_t size = READ_REG(sizeReg);
+size_t stride = READ_REG(strideReg);
 
 /*
 // For debug, print the indexes of each register as well as their contents
